Give the 'header logo' text it's own div
Previously the 'header logo' text wasnt dived off, so to style the text
the 'header' div had to be styled causing the typography based
styles to be inherited by the elements children ('ul' element).

Giving the 'header logo' text it's own div allows styles to be applied
to it individually without effecting the other surrounding elements.